chore: add tests for sanic and aiohttp jinja graphiql render
