+++ tests/aiohttp/app.py
@@ -3,7 +3,8 @@ from urllib.parse import urlencode
 from aiohttp import web
 
 from graphql_server.aiohttp import GraphQLView
-from tests.aiohttp.schema import Schema
+
+from .schema import Schema
 
 
 def create_app(schema=Schema, **kwargs):
@@ -13,10 +14,5 @@ def create_app(schema=Schema, **kwargs):
     return app
 
 
-def url_string(**url_params):
-    base_url = "/graphql"
-
-    if url_params:
-        return f"{base_url}?{urlencode(url_params)}"
-
-    return base_url
+def url_string(url="/graphql", **url_params):
+    return f"{url}?{urlencode(url_params)}" if url_params else url
