+++ tests/aiohttp/test_graphiqlview.py
@@ -1,37 +1,27 @@
 import pytest
-import pytest_asyncio
-from aiohttp.test_utils import TestClient, TestServer
 from jinja2 import Environment
 
-from tests.aiohttp.app import create_app, url_string
-from tests.aiohttp.schema import AsyncSchema, Schema, SyncSchema
+from .app import create_app, url_string
+from .schema import AsyncSchema, SyncSchema
 
 
-@pytest.fixture
-def app():
-    app = create_app()
-    return app
+@pytest.mark.asyncio
+@pytest.mark.parametrize(
+    "app",
+    [
+        create_app(graphiql=True),
+        create_app(graphiql=True, jinja_env=Environment()),
+        create_app(graphiql=True, jinja_env=Environment(enable_async=True)),
+    ],
+)
+async def test_graphiql_is_enabled(app, client):
+    response = await client.get(
+        url_string(query="{test}"),
+        headers={"Accept": "text/html"},
+    )
+    assert response.status == 200
 
-
-@pytest_asyncio.fixture
-async def client(app):
-    client = TestClient(TestServer(app))
-    await client.start_server()
-    yield client
-    await client.close()
-
-
-@pytest.fixture
-def view_kwargs():
-    return {
-        "schema": Schema,
-        "graphiql": True,
-    }
-
-
-@pytest.fixture
-def pretty_response():
-    return (
+    pretty_response = (
         "{\n"
         '  "data": {\n'
         '    "test": "Hello World"\n'
@@ -39,39 +29,7 @@ def pretty_response():
         "}".replace('"', '\\"').replace("\n", "\\n")
     )
 
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize("app", [create_app(graphiql=True)])
-async def test_graphiql_is_enabled(app, client):
-    response = await client.get(
-        url_string(query="{test}"), headers={"Accept": "text/html"}
-    )
-    assert response.status == 200
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize("app", [create_app(graphiql=True)])
-async def test_graphiql_simple_renderer(app, client, pretty_response):
-    response = await client.get(
-        url_string(query="{test}"),
-        headers={"Accept": "text/html"},
-    )
-    assert response.status == 200
     assert pretty_response in await response.text()
-
-
-class TestJinjaEnv:
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize(
-        "app", [create_app(graphiql=True, jinja_env=Environment(enable_async=True))]
-    )
-    async def test_graphiql_jinja_renderer_async(self, app, client, pretty_response):
-        response = await client.get(
-            url_string(query="{test}"),
-            headers={"Accept": "text/html"},
-        )
-        assert response.status == 200
-        assert pretty_response in await response.text()
 
 
 @pytest.mark.asyncio
@@ -84,7 +42,10 @@ async def test_graphiql_html_is_not_accepted(client):
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("app", [create_app(graphiql=True)])
+@pytest.mark.parametrize(
+    "app",
+    [create_app(graphiql=True), create_app(graphiql=True, jinja_env=Environment())],
+)
 async def test_graphiql_get_mutation(app, client):
     response = await client.get(
         url_string(query="mutation TestMutation { writeTest { test } }"),
@@ -95,7 +56,10 @@ async def test_graphiql_get_mutation(app, client):
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("app", [create_app(graphiql=True)])
+@pytest.mark.parametrize(
+    "app",
+    [create_app(graphiql=True), create_app(graphiql=True, jinja_env=Environment())],
+)
 async def test_graphiql_get_subscriptions(app, client):
     response = await client.get(
         url_string(
@@ -109,7 +73,16 @@ async def test_graphiql_get_subscriptions(app, client):
 
 @pytest.mark.asyncio
 @pytest.mark.parametrize(
-    "app", [create_app(schema=AsyncSchema, enable_async=True, graphiql=True)]
+    "app",
+    [
+        create_app(schema=AsyncSchema, enable_async=True, graphiql=True),
+        create_app(
+            schema=AsyncSchema,
+            enable_async=True,
+            graphiql=True,
+            jinja_env=Environment(),
+        ),
+    ],
 )
 async def test_graphiql_enabled_async_schema(app, client):
     response = await client.get(
@@ -136,7 +109,13 @@ async def test_graphiql_enabled_async_schema(app, client):
 
 @pytest.mark.asyncio
 @pytest.mark.parametrize(
-    "app", [create_app(schema=SyncSchema, enable_async=True, graphiql=True)]
+    "app",
+    [
+        create_app(schema=SyncSchema, enable_async=True, graphiql=True),
+        create_app(
+            schema=SyncSchema, enable_async=True, graphiql=True, jinja_env=Environment()
+        ),
+    ],
 )
 async def test_graphiql_enabled_sync_schema(app, client):
     response = await client.get(
