+++ tests/sanic/test_graphiqlview.py
@@ -5,9 +5,22 @@ from .app import create_app, url_string
 from .schema import AsyncSchema, SyncSchema
 
 
-@pytest.fixture
-def pretty_response():
-    return (
+@pytest.mark.parametrize(
+    "app",
+    [
+        create_app(graphiql=True),
+        create_app(graphiql=True, jinja_env=Environment()),
+        create_app(graphiql=True, jinja_env=Environment(enable_async=True)),
+    ],
+)
+def test_graphiql_is_enabled(app):
+    _, response = app.test_client.get(
+        uri=url_string(query="{test}"), headers={"Accept": "text/html"}
+    )
+
+    assert response.status == 200
+
+    pretty_response = (
         "{\n"
         '  "data": {\n'
         '    "test": "Hello World"\n'
@@ -15,41 +28,6 @@ def pretty_response():
         "}".replace('"', '\\"').replace("\n", "\\n")
     )
 
-
-@pytest.mark.parametrize("app", [create_app(graphiql=True)])
-def test_graphiql_is_enabled(app):
-    _, response = app.test_client.get(
-        uri=url_string(query="{test}"), headers={"Accept": "text/html"}
-    )
-    assert response.status == 200
-
-
-@pytest.mark.parametrize("app", [create_app(graphiql=True)])
-def test_graphiql_simple_renderer(app, pretty_response):
-    _, response = app.test_client.get(
-        uri=url_string(query="{test}"), headers={"Accept": "text/html"}
-    )
-    assert response.status == 200
-    assert pretty_response in response.body.decode("utf-8")
-
-
-@pytest.mark.parametrize("app", [create_app(graphiql=True, jinja_env=Environment())])
-def test_graphiql_jinja_renderer(app, pretty_response):
-    _, response = app.test_client.get(
-        uri=url_string(query="{test}"), headers={"Accept": "text/html"}
-    )
-    assert response.status == 200
-    assert pretty_response in response.body.decode("utf-8")
-
-
-@pytest.mark.parametrize(
-    "app", [create_app(graphiql=True, jinja_env=Environment(enable_async=True))]
-)
-def test_graphiql_jinja_async_renderer(app, pretty_response):
-    _, response = app.test_client.get(
-        uri=url_string(query="{test}"), headers={"Accept": "text/html"}
-    )
-    assert response.status == 200
     assert pretty_response in response.body.decode("utf-8")
