+++ tests/sanic/test_graphqlview.py
@@ -20,7 +20,6 @@ def json_dump_kwarg_list(**kwargs):
     return json.dumps([kwargs])
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_allows_get_with_query_param(app):
     _, response = app.test_client.get(uri=url_string(query="{test}"))
 
@@ -28,7 +27,6 @@ def test_allows_get_with_query_param(app):
     assert response_json(response) == {"data": {"test": "Hello World"}}
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_allows_get_with_variable_values(app):
     _, response = app.test_client.get(
         uri=url_string(
@@ -41,7 +39,6 @@ def test_allows_get_with_variable_values(app):
     assert response_json(response) == {"data": {"test": "Hello Dolly"}}
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_allows_get_with_operation_name(app):
     _, response = app.test_client.get(
         uri=url_string(
@@ -63,7 +60,6 @@ def test_allows_get_with_operation_name(app):
     }
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_reports_validation_errors(app):
     _, response = app.test_client.get(
         uri=url_string(query="{ test, unknownOne, unknownTwo }")
@@ -84,7 +80,6 @@ def test_reports_validation_errors(app):
     }
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_errors_when_missing_operation_name(app):
     _, response = app.test_client.get(
         uri=url_string(
@@ -106,7 +101,6 @@ def test_errors_when_missing_operation_name(app):
     }
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_errors_when_sending_a_mutation_via_get(app):
     _, response = app.test_client.get(
         uri=url_string(
@@ -125,7 +119,6 @@ def test_errors_when_sending_a_mutation_via_get(app):
     }
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_errors_when_selecting_a_mutation_within_a_get(app):
     _, response = app.test_client.get(
         uri=url_string(
@@ -147,7 +140,6 @@ def test_errors_when_selecting_a_mutation_within_a_get(app):
     }
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_allows_mutation_to_exist_within_a_get(app):
     _, response = app.test_client.get(
         uri=url_string(
@@ -163,7 +155,6 @@ def test_allows_mutation_to_exist_within_a_get(app):
     assert response_json(response) == {"data": {"test": "Hello World"}}
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_allows_post_with_json_encoding(app):
     _, response = app.test_client.post(
         uri=url_string(),
@@ -175,7 +166,6 @@ def test_allows_post_with_json_encoding(app):
     assert response_json(response) == {"data": {"test": "Hello World"}}
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_allows_sending_a_mutation_via_post(app):
     _, response = app.test_client.post(
         uri=url_string(),
@@ -187,7 +177,6 @@ def test_allows_sending_a_mutation_via_post(app):
     assert response_json(response) == {"data": {"writeTest": {"test": "Hello World"}}}
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_allows_post_with_url_encoding(app):
     # Example of how sanic does send data using url enconding
     # can be found at their repo.
@@ -203,7 +192,6 @@ def test_allows_post_with_url_encoding(app):
     assert response_json(response) == {"data": {"test": "Hello World"}}
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_supports_post_json_query_with_string_variables(app):
     _, response = app.test_client.post(
         uri=url_string(),
@@ -218,7 +206,6 @@ def test_supports_post_json_query_with_string_variables(app):
     assert response_json(response) == {"data": {"test": "Hello Dolly"}}
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_supports_post_json_query_with_json_variables(app):
     _, response = app.test_client.post(
         uri=url_string(),
@@ -233,7 +220,6 @@ def test_supports_post_json_query_with_json_variables(app):
     assert response_json(response) == {"data": {"test": "Hello Dolly"}}
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_supports_post_url_encoded_query_with_string_variables(app):
     _, response = app.test_client.post(
         uri=url_string(),
@@ -250,7 +236,6 @@ def test_supports_post_url_encoded_query_with_string_variables(app):
     assert response_json(response) == {"data": {"test": "Hello Dolly"}}
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_supports_post_json_query_with_get_variable_values(app):
     _, response = app.test_client.post(
         uri=url_string(variables=json.dumps({"who": "Dolly"})),
@@ -264,7 +249,6 @@ def test_supports_post_json_query_with_get_variable_values(app):
     assert response_json(response) == {"data": {"test": "Hello Dolly"}}
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_post_url_encoded_query_with_get_variable_values(app):
     _, response = app.test_client.post(
         uri=url_string(variables=json.dumps({"who": "Dolly"})),
@@ -280,7 +264,6 @@ def test_post_url_encoded_query_with_get_variable_values(app):
     assert response_json(response) == {"data": {"test": "Hello Dolly"}}
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_supports_post_raw_text_query_with_get_variable_values(app):
     _, response = app.test_client.post(
         uri=url_string(variables=json.dumps({"who": "Dolly"})),
@@ -292,7 +275,6 @@ def test_supports_post_raw_text_query_with_get_variable_values(app):
     assert response_json(response) == {"data": {"test": "Hello Dolly"}}
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_allows_post_with_operation_name(app):
     _, response = app.test_client.post(
         uri=url_string(),
@@ -316,7 +298,6 @@ def test_allows_post_with_operation_name(app):
     }
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_allows_post_with_get_operation_name(app):
     _, response = app.test_client.post(
         uri=url_string(operationName="helloWorld"),
@@ -353,7 +334,6 @@ def test_not_pretty_by_default(app):
     assert response.body.decode() == '{"data":{"test":"Hello World"}}'
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_supports_pretty_printing_by_request(app):
     _, response = app.test_client.get(uri=url_string(query="{test}", pretty="1"))
 
@@ -362,7 +342,6 @@ def test_supports_pretty_printing_by_request(app):
     )
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_handles_field_errors_caught_by_graphql(app):
     _, response = app.test_client.get(uri=url_string(query="{thrower}"))
     assert response.status == 200
@@ -378,7 +357,6 @@ def test_handles_field_errors_caught_by_graphql(app):
     }
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_handles_syntax_errors_caught_by_graphql(app):
     _, response = app.test_client.get(uri=url_string(query="syntaxerror"))
     assert response.status == 400
@@ -392,7 +370,6 @@ def test_handles_syntax_errors_caught_by_graphql(app):
     }
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_handles_errors_caused_by_a_lack_of_query(app):
     _, response = app.test_client.get(uri=url_string())
 
@@ -402,7 +379,6 @@ def test_handles_errors_caused_by_a_lack_of_query(app):
     }
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_handles_batch_correctly_if_is_disabled(app):
     _, response = app.test_client.post(
         uri=url_string(), content="[]", headers={"content-type": "application/json"}
@@ -418,7 +394,6 @@ def test_handles_batch_correctly_if_is_disabled(app):
     }
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_handles_incomplete_json_bodies(app):
     _, response = app.test_client.post(
         uri=url_string(),
@@ -432,7 +407,6 @@ def test_handles_incomplete_json_bodies(app):
     }
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_handles_plain_post_text(app):
     _, response = app.test_client.post(
         uri=url_string(variables=json.dumps({"who": "Dolly"})),
@@ -445,7 +419,6 @@ def test_handles_plain_post_text(app):
     }
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_handles_poorly_formed_variables(app):
     _, response = app.test_client.get(
         uri=url_string(
@@ -458,7 +431,6 @@ def test_handles_poorly_formed_variables(app):
     }
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_handles_unsupported_http_methods(app):
     _, response = app.test_client.put(uri=url_string(query="{test}"))
     assert response.status == 405
@@ -475,7 +447,6 @@ def test_handles_unsupported_http_methods(app):
     }
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_passes_request_into_request_context(app):
     _, response = app.test_client.get(uri=url_string(query="{request}", q="testing"))
 
@@ -513,7 +484,6 @@ def test_context_remapped_if_not_mapping(app):
     assert "Request" in res["data"]["context"]["request"]
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_post_multipart_data(app):
     query = "mutation TestMutation { writeTest { test } }"
 
@@ -603,7 +573,6 @@ def test_async_schema(app):
     assert response_json(response) == {"data": {"a": "hey", "b": "hey2", "c": "hey3"}}
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_preflight_request(app):
     _, response = app.test_client.options(
         uri=url_string(), headers={"Access-Control-Request-Method": "POST"}
@@ -612,7 +581,6 @@ def test_preflight_request(app):
     assert response.status == 200
 
 
-@pytest.mark.parametrize("app", [create_app()])
 def test_preflight_incorrect_request(app):
     _, response = app.test_client.options(
         uri=url_string(), headers={"Access-Control-Request-Method": "OPTIONS"}
